requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
pyarrow>=12.0.0
openai>=1.0.0
python-dotenv>=1.0.0
openpyxl>=3.1.0